requirements.txt
Flask==2.3.3
flask-smorest==0.41.0
flask-swagger-ui==5.21.0
marshmallow==3.21.1
requests==2.32.5
beautifulsoup4==4.14.3
PyJWT==2.11.0
mysql-connector-python>=9.6.0